% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scales.R
\name{add1.item.scale}
\alias{add1.item.scale}
\title{Iteratively add items to a scale and recompute scale measures}
\usage{
\method{add1}{item.scale}(x, scope)
}
\arguments{
\item{x}{An \code{item.scale}}

\item{scope}{A character vector containing the list of items that will be added.  Items
must be present either in the original data from of {x} or in the environment.}
}
\value{
An \code{n} by 3 matrix where \code{n} is the number of items to be added.
  Includes the mean, standard deviation, and alpha of each new scale as items are added.
}
\description{
This allows you to take an item.scale and iteratively add new items.  It recalculates the scale and
also calculates new Cronbach's alpha for each new scale.
}
\details{
This is useful during exploratory analysis in determining which items are needed in a scale.
Each item in \code{scope} will be added separately in the order they are given and a
new scale/alpha will be computed.
}

