% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_types.R
\name{add_attention_check}
\alias{add_attention_check}
\title{Attention Check questions}
\usage{
add_attention_check(q_name, valid)
}
\arguments{
\item{q_name}{The column name of the attention check question}

\item{valid}{What value is to be considered valid}
}
\description{
Tells the question auto-detection system that a specific question is an attention check question
}
\details{
Attention check questions are questions whose goal is to determine whether the subject is
paying close attention to questions.  It has exactly one right answer, and any other answer
is invalid.
}
