% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_answer.R
\name{add_multiple_answer}
\alias{add_multiple_answer}
\title{Add a new set of multiple answer choice options.  Mulitple answers have checkboxes that
allow the survey participant to choose more than one answer.}
\usage{
add_multiple_answer(choice_set, case_sensitive = F)
}
\arguments{
\item{choice_set}{Character vector of options}

\item{case_sensitive}{Are the options case sensitive?}
}
\description{
Add a new set of multiple answer choice options.  Mulitple answers have checkboxes that
allow the survey participant to choose more than one answer.
}
