% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_choice.R
\name{add_multiple_choice}
\alias{add_multiple_choice}
\title{Add a new set of multiple choice options}
\usage{
add_multiple_choice(choice_set, ordered = T, case_sensitive = F)
}
\arguments{
\item{choice_set}{Character vector of options}

\item{ordered}{Are the options in a particular order?}

\item{case_sensitive}{Are the options case sensitive?}
}
\description{
Add a new set of multiple choice options
}
