% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{add_perc}
\alias{add_perc}
\title{Add percentage column}
\usage{
add_perc(
  x,
  count_col = NULL,
  .total = NULL,
  name = "perc",
  margins = NULL,
  format = TRUE
)
}
\arguments{
\item{x}{The dataset}

\item{count_col}{The name of the column (tidy-select) that contains which column should be counted}

\item{.total}{(optional) The name of the column containing totals}

\item{name}{The name of the new column being added with the percentage (default: "perc")}

\item{format}{}
}
\description{
Adds a percentage column based on the counts in the count_col.  Automatically quoted.
.total is to specify a total count (e.g. in chooseany columns, where the sum is not the same as the total N)
name is the name of the column to add
margins is which columns to add margins for (tidy-select)
}
