% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detector.R
\name{add_question_type}
\alias{add_question_type}
\title{New question type.}
\usage{
add_question_type(name, detector, processor)
}
\arguments{
\item{name}{Text string containing the name of this question type}

\item{detector}{Function that takes a single argument -- a vector of data -- and
returns True or False -- whether this data is of the correct type}

\item{processor}{Function that takes a single argument -- a vector of data --
processes it, and returns the processed data}
}
\description{
Adds a new type of question to the survey question auto-detector.
}
