% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scales.R
\name{alt_alpha}
\alias{alt_alpha}
\title{Alternative way of calculating cronbach's alpha}
\usage{
alt_alpha(v1)
}
\arguments{
\item{v1}{A matrix or data frame where each column is an item, and each row is a data point.}
}
\value{
A list containing three items: sample.size, number.of.items, and alpha
}
\description{
Cronbach's alpha is a measure of internal reliability.
}
\details{
This is a slightly different way to calculate it than what is in the psy package.

Note: Data is run through na.omit to remove NAs when calculating alpha.
}

