% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/surveys.R
\name{apply_columns}
\alias{apply_columns}
\title{Apply a function to specific columns}
\usage{
apply_columns(df, fun, cols, ...)
}
\arguments{
\item{df}{The data.frame to apply to}

\item{fun}{The function to apply.  The first parameter should be the column of data}

\item{cols}{The set of columns to apply to}

\item{...}{Additional parameters to pass into \code{fun}}
}
\value{
A new copy of the data frame with the modified columns
}
\description{
Apply a function to a specified set of columns in a data frame
}
\examples{
df <- data.frame(one=c("1.0", "2.0"), two=c("3.0", "4.0"))
df <- apply_columns(df, as.numeric, c("one", "two"))
}

