% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detector.R
\name{detect.question}
\alias{detect.question}
\title{Question Auto-Detect}
\usage{
detect.question(column, col_name = "")
}
\arguments{
\item{column}{A single vector of data from a survey}

\item{col_name}{(optional) The name of the column}
}
\value{
A vector of processed data
}
\description{
Automatically detect what type of question this column of data came from, and process it
}
