% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detector.R
\name{detect.survey}
\alias{detect.survey}
\title{Survey Cleaning Auto-detect}
\usage{
detect.survey(frame)
}
\arguments{
\item{frame}{The data.frame containing the survey}
}
\value{
A new data.frame containing the processed survey
}
\description{
Automatically clean a survey by iterating through a data frame and automatically detecting
each question's type and processing it
}

