% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scales.R
\name{drop1.item.scale}
\alias{drop1.item.scale}
\title{Iteratively drop items from a scale and recompute scale measures}
\usage{
\method{drop1}{item.scale}(x, scope)
}
\arguments{
\item{x}{An \code{item.scale}}

\item{scope}{A character vector containing the list of items that will be dropped.}
}
\value{
An \code{n} by 3 matrix where \code{n} is the number of items to be dropped.
  Includes the mean, standard deviation, and alpha of each new scale as items are dropped.
}
\description{
This allows you to take an item.scale and iteratively drop items.  It recalculates the scale and
also calculates new Cronbach's alpha for each new scale.
}
\details{
This is useful during exploratory analysis in determining whether all items are needed in a scale.
Each item in \code{scope} will be dropped separately in the order they are given and a
new scale/alpha will be computed.

If scope is omitted, all items in the scale will be iteratively dropped.
}

