% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{factanal.item.scale}
\alias{factanal.item.scale}
\title{Conduct a factor analysis on the items of an \code{item.scale}}
\usage{
\method{factanal}{item.scale}(x, factors, ...)
}
\arguments{
\item{x}{An \code{item.scale} that contains the list of items to factor analyze}

\item{factors}{the number of factors to extract}
}
\value{
An object of class \code{factanal}
}
\description{
Extract the list of items in an \code{item.scale} and then conduct a maximum likelihood
factor analysis on that complete list of items
}
