% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorsets.R
\name{factorset}
\alias{factorset}
\title{Identify a set of candidate questions for new factors.}
\usage{
factorset(scale, factors = 4, questions = 5, cut = 0.4)
}
\arguments{
\item{scale}{An \code{item.scale} object that specifies the set of questions to be analyzed}

\item{factors}{The number of factors to extract}

\item{questions}{The maximum number of questions to identify in each factor.}

\item{cut}{The minimum factor loading for a question to be considered part of a factor}
}
\value{
A \code{factorset} object
}
\description{
Use maximum likelihood factor analysis to identify a set of candidate questions for new factors.
}
\details{
Uses \code{factanal} to conduct a factor analysis of a large number of questions.  Then identifies the
top Q questions in each of F factors.  It reruns a new factor analysis on just those questions to ensure
they still hold together
}
