% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_types.R
\name{ignore_question}
\alias{ignore_question}
\title{Ignore a specific question}
\usage{
ignore_question(q_name)
}
\arguments{
\item{q_name}{The name of the question to ignore}
}
\description{
Tells the question auto-detector to ignore a specific question.  This doesn't do any processing
just passes the raw character vector through
}
