% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scales.R
\name{item.scale}
\alias{item.scale}
\title{Functions to form a scale from a set of survey items}
\usage{
item.scale(formula, data = NULL, subset = NULL, na.action = na.pass,
  drop.unused.levels = F, scale.fun = mean, standard = FALSE,
  numeric = FALSE, ...)
}
\arguments{
\item{formula}{A formula specifying the items to combine into a scale
Should be of the form \code{~ item1 + item2 + item3}}

\item{data}{The data frame to use to get the data from}

\item{subset}{Specifies if only a specific subset of the data should be used}

\item{na.action}{What should be done with NAs?}

\item{scale.fun}{which function to compose the scale.   examples include sum, mean}

\item{standard}{should the individual items be standardized before inclusion in the scale?}

\item{numeric}{should the individual items be cast into numeric before inclusion in the scale?}

\item{...}{currently unused}

\item{drop.unused.level}{If factors are being combined, should we only include levels that
are actually used?}
}
\value{
An object of type \code{item.scale}.  This is fundamentally a vector, and can
  be used as such.  It has a number of additional attributes (such as \code{cronbach})
  that are useful in future calculations and functions
}
\description{
Combines multiple items, specified using a formula interface, into a single "scale"
}

