% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_types.R
\name{known_question}
\alias{known_question}
\title{Known questions}
\usage{
known_question(q_name, levels, options = NULL, ordered = F)
}
\arguments{
\item{q_name}{The column name of the attention check question}

\item{levels}{What the different options should be called}

\item{options}{What the options look like in the original data.  NULL == autodetect}

\item{ordered}{Should the resulting factor be ordered?}
}
\description{
Tells the question auto-detection system that a specific question is known with a predefined set of responses
}
