% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detector.R
\name{load_survey}
\alias{load_survey}
\title{Load a survey dataset from a file}
\usage{
load_survey(file, ...)
}
\arguments{
\item{file}{Name of the file to load}
}
\value{
A data.frame containing the processed dataset
}
\description{
Loads a survey dataset from a file, auto-detects all of the columns, and then processes the
dataset
}
