% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/surveys.R
\name{make_columns_numeric}
\alias{make_columns_numeric}
\title{Add numeric versions of columns}
\usage{
make_columns_numeric(df, cols, fun = as.numeric, ...)
}
\arguments{
\item{df}{The data.frame with the source data}

\item{cols}{A list of column names to convert}

\item{fun}{The function to apply to all of the columns}

\item{...}{Additional parameters to fun}
}
\value{
A new data frame with additional columns.  Each column listed in \code{cols} will have
  a new version added, column\code{_n}, containing the data data as in column convert to numeric
}
\description{
Take a data frame and add numeric versions of specified factor columns
}

