% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/surveys.R
\name{mf}
\alias{mf}
\title{Create a formula from a list of variable names}
\usage{
mf(..., type = "+")
}
\arguments{
\item{...}{One or more lists of variable names to make into a formula}

\item{type}{Combination method.  Defaults to '+'.  It might also be useful to do '*'}
}
\value{
A formula containing all of the specified names
}
\description{
Creates a formula from a list of variable names.  If more than one list of names, it combines
them using \code{\link{combine_names}} to create all combinations of names.
}
\examples{
mf(c("var"), c("one", "two"), "_n")
}

