% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot objects}

\item{plotlist}{A list of ggplot objects.  Will be combined with objects in \code{...}}

\item{cols:}{Number of columns in layout}

\item{layout:}{A matrix specifying the layout. If present, 'cols' is ignored.}
}
\value{
A single grid object containing all of the plots arranged as specified in \code{cols} or \code{layout}
}
\description{
ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)
}
