% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{ns}
\alias{ns}
\title{Select names from a tibble, and returns a character vector}
\usage{
ns(data, ...)
}
\arguments{
\item{data}{The dataset (data frame or tibble)}

\item{...}{tidy-select name selection}
}
\value{
Character vector of column names
}
\description{
Select names from a tibble, and returns a character vector
}
