% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/factorsets.R
\name{plot.factorset}
\alias{plot.factorset}
\title{Draw a plot of a factorset}
\usage{
\method{plot}{factorset}(fs, ymax = 1150)
}
\arguments{
\item{fs}{A \code{factorset} object}

\item{ymax}{The maximum y value to draw.  This ensures that all items are visually comparable.}
}
\description{
Draws a set of histograms where each question has its own histogram, and each factor is a row
}

