% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.item.scale}
\alias{plot.item.scale}
\title{Generate a set of histograms for the items in an \code{item.scale}}
\usage{
\method{plot}{item.scale}(s, cols = NULL, ...)
}
\arguments{
\item{s}{An \code{item.scale} object}

\item{cols}{(optional) The number of columns in the resulting plot.  If omitted, the function will make its best guess}

\item{...}{Additional parameters passed to \code{plot_question}.  Commonly used to set a \code{ymax} value
so all plots have the same y range}
}
\value{
A grid object containing the set of histograph ggplot2 plots
}
\description{
Extracts the original data from an \code{item.scale}, generates a histogram for each question, and
plots them side-by-side
}
