% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_list}
\alias{plot_list}
\title{Generate a series of histograms for all of the questions in an \code{item.scale}}
\usage{
plot_list(s, ...)
}
\arguments{
\item{s}{An \code{item.scale} object}

\item{...}{Any additional parameters for \code{plot_question}}
}
\value{
A \code{list} of ggplot2 objects
}
\description{
This will iteratively generate a list of histogram plots, one for each of the questions listed in an \code{item.scale}
}
