% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_question}
\alias{plot_question}
\title{Generate a histogram of the responses to a single survey question}
\usage{
plot_question(d, question, labels = NULL, ymax = 1200)
}
\arguments{
\item{d}{A data frame containing the source data}

\item{question}{Which column in \code{d} to plot}

\item{labels}{an optional list of labels for the question. This is usually the list of
question options from the survey}

\item{ymax}{The maximum y value (ylim) for the plot.  This is used to ensure that a set of plots are comparable.}
}
\value{
A ggplot2 plot object containing a histogram
}
\description{
Using ggplot2, generate a plot object that will display a histogram of the responses to
single survey item.
}

