% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/factorsets.R
\name{print.factorset}
\alias{print.factorset}
\title{Print a summary of a factorset}
\usage{
\method{print}{factorset}(fs, cut = NULL)
}
\arguments{
\item{fs}{The \code{factorset} object to be printed}

\item{cut}{(optional) The cutoff for \code{print.factanal}; Only prints factor loading above this value}
}
\description{
Prints out the final factor analysis of a factorset
}

