% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scales.R
\name{print.item.scale}
\alias{print.item.scale}
\title{Prints information about an \code{item.scale} object}
\usage{
\method{print}{item.scale}(x, digits = 3)
}
\arguments{
\item{x}{An \code{item.scale} object}

\item{digits}{The number of digits to print when printing numbers.  Passed to \code{format()}}
}
\description{
Outputs standard information found in an \code{item.scale}, such as number of items, number of
responses, scale mean, and Cronbach's alpha
}

