% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{print.summary.item.scale}
\alias{print.summary.item.scale}
\title{Prints out long summary information about an \code{item.scale}}
\usage{
\method{print}{summary.item.scale}(x, digits = 3)
}
\arguments{
\item{x}{An \code{item.scale} object}

\item{digits}{The number of digits to print when printing numbers.  Passed to \code{format()}}
}
\description{
Prints out basic summary information, such as mean, sd, and cronbach's alpha.  Also prints out long
summary information, such as the mean and sd of each item, and the item-item correlations
}
