% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detector.R
\name{remove_question_type}
\alias{remove_question_type}
\title{Remove a question type from the auto-detector}
\usage{
remove_question_type(name)
}
\arguments{
\item{name}{Character string name of the question type to remove}
}
\description{
Remove a question type from the auto-detector
}

