% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detector.R
\name{reset_surveys}
\alias{reset_surveys}
\title{Resets the surveys package}
\usage{
reset_surveys()
}
\description{
Removes all known question types and answer types, and resets the surveys packag
back to where it is when it is first loaded
}
