% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scales.R
\name{standardize}
\alias{standardize}
\title{Function to standardize an item or scale.}
\usage{
standardize(x)
}
\arguments{
\item{x}{A vector of data values to be standardized}
}
\value{
Standardized data values
}
\description{
Standardizes a vector or column of data by centering around the mean and making
  standard deviation equal to one
}
\details{
Note: This function automatically removes NAs when calculating the mean and standard deviation
}
\examples{
ex <- standardize(rnorm(100, 10, 5))
mean(ex)
sd(ex)
}

