% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarize_chooseany}
\alias{summarize_chooseany}
\title{Summarize a chooseany question}
\usage{
summarize_chooseany(data, question, width = default_width, ...)
}
\arguments{
\item{data}{Data frame or tibble with the dataset}

\item{question}{Which survey question, as a string.}

\item{width}{Truncates question names that are too long to width

Since this attempts to summarize a choose-any-of-the-options style question
(so multiple answers are possible), the function assumes that each answer is
a separate column, and that all columns start with the same string (the name
of the question), are followed by an underscore, followed by the value of the
option.   The function uses starts_with() to identify the appropraite set of
columns to summarize (which is why question should be a string)}
}
\description{
Summarize a chooseany question
}
