% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarize_chooseone}
\alias{summarize_chooseone}
\title{Summarizes a multiple choice style question}
\usage{
summarize_chooseone(data, question, width = default_width, ...)
}
\arguments{
\item{data}{Data frame or tibble with the dataset}

\item{question}{Which survey question, as a string.}

\item{width}{Truncates question names that are too long to width}
}
\description{
Summarizes a multiple choice style question
}
