% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scales.R
\name{summary.item.scale}
\alias{summary.item.scale}
\title{Calculates additional summary data about an \code{item.scale}}
\usage{
\method{summary}{item.scale}(x, maxsum = NULL, digits = NULL)
}
\arguments{
\item{x}{An \code{item.scale} object}

\item{maxsum}{Not sure what this is}

\item{digits}{Unused}
}
\value{
An object of class \code{summary.item.scale}
}
\description{
Calculates summary data such as individual item means and standard deviations
}

