% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/surveys.R
\name{combine_names}
\alias{combine_names}
\title{Create all combinations of sets of strings}
\usage{
combine_names(..., sep = ".")
}
\arguments{
\item{...}{One or more character vectors}

\item{sep}{String to use as a separator character}
}
\value{
A vector of all possible combinations of the input paramters
}
\description{
Combine two sets of strings to create all possible combinations
}
\examples{
combine_names(c("is", "as"), c("numeric", "logical"))
}

